// apps/back-office/marketplace-fe/src/enums/rfx-status.ts
export enum ERfxStatus {
  DRAFT = 'DRAFT',
  TEAM_REVIEWAL = 'TEAM_REVIEWAL',
  ADJUSTMENT = 'ADJUSTMENT',

  SUBMITTED = 'SUBMITTED',
  SUBMITTED_EVALUATION = 'SUBMITTED_EVALUATION',

  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  CANCELLED = 'CANCELLED',

  EVALUATION = 'EVALUATION',
  EVALUATION_REJECTED = 'EVALUATION_REJECTED',

  AUCTION = 'AUCTION',
  ENDED = 'ENDED',
}